fix(sql): resolve GetMonthReportRow fields to int64 with bigint casting
Cast to BIGINT wherever possible and necessary to get sqlc to generate the rows of GetMonthReportRow as int64 types again

+++ sql/schema/006_assignment.sql
@@ -59,7 +59,7 @@ SELECT
     category_id,
     assigned,
     activity,
-    SUM(assigned + activity) OVER (PARTITION BY category_id ORDER BY month) AS balance
+    (SUM(assigned + activity) OVER (PARTITION BY category_id ORDER BY month))::bigint AS balance
 FROM report
 ORDER BY month, category_name;
